Generate stubs for `DatabindingMapperImpl` so that we can completely disable header compilation for databinding targets.
Header compilation is nice but that also means annotation processors are run twice per module. Databinding is already slow + consumer transitive R.txts which affects parallelism. This change generates `DatabindingMapperImpl` for a module so that that public ABI of a module is fully complete (we already generate other stubs).

This way by setting generates_api to `False` in bazel, we can avoid running `ProcessDatabinding` twice.